fitness as float-datatype # some parameters changes ## added bucket rewarding and random covering

+++ PacMan/src/game/player/ghost/group4/IClassifierGenerator.java
@@ -4,6 +4,6 @@ import game.player.ghost.group4.system.ZCSEntry;
 
 public interface IClassifierGenerator {
 
-	ZCSEntry generateRandomClassifierForObservation(int observation, int fitness);
+	ZCSEntry generateRandomClassifierForObservation(int observation, float fitness);
 	ZCSEntry generateGeneticClassifier(int observation, ZCSEntry a, ZCSEntry b);
 }
